if style-src CSP directive, supply nonce to script
If the client website has defined a Content Security Policy that include
'style-src', then we need to apply a nonce to the Betterlink script --
which will apply the nonce to each style element that is created by the
script.

The nonce is a signal to the browser that the style element is allowed.
If a nonce is already supplied by the server, then we should use that
nonce. If one is not provided, then we should create it ourselves.