don't inject on major problem sites
In general, we don't want to inject Betterlink on pages where the
content is personalized (Facebook) or private to a user (email).

The best protection is to update the manifest to exclude those domains.
However:
a) we will not be able to maintain an exhaustive list
b) the structure of the manifest doesn't allow refined filters

For example, the help pages on Facebook are at www.facebook.com/help --
but the personalized content is under the same domain. Doing the broad
filtering is a better solution because it still allows the client
website to include the betterlink.js script on their own to the pages
where it makes sense.